ElectricityPricePrediction/requirements.txt
pandas
numpy
matplotlib
seaborn
scikit-learn
xgboost
tensorflow
optuna
openmeteo-requests
requests-cache
retry_requests
